src/e-paper-driver/requirements.txt
Pillow>=10.0
spidev>=3.6
gpiozero>=2.0
